Move deleteExpiredTtl method to ChatManager
